+++ sdk/pennylane/quantum_api_pennylane/__init__.py
@@ -1,0 +1,3 @@
+from quantum_api_pennylane.device import QuantumApiDevice
+
+__all__ = ["QuantumApiDevice"]